+++ src/NXO.Functions/Program.cs
@@ -4,11 +4,15 @@ using Microsoft.Azure.Functions.Worker.Builder;
 using Microsoft.Azure.Functions.Worker.OpenTelemetry;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using NXO.Functions.Services;
 using OpenTelemetry;
 
 var builder = FunctionsApplication.CreateBuilder(args);
 
 builder.ConfigureFunctionsWebApplication();
+
+// Registrer services
+builder.Services.AddSingleton<EmailClassifier>();
 
 // TODO: Tilføj Azure Monitor connection string for produktion
 // builder.Services.AddOpenTelemetry()